Adding validation schema for incoming requests

+++ src/controllers/task.controller.js
@@ -1,5 +1,6 @@
 const httpStatus = require("http-status");
 const { isEmpty } = require("lodash");
+const { default: mongoose } = require("mongoose");
 
 const { roles } = require("@src/lib/constant");
 const { catchAsync, generateId, pick } = require("@src/lib/utils");
@@ -36,7 +37,27 @@ const getTaskWithComments = catchAsync(async (req, res, next) => {
 })
 
 const deleteTask = catchAsync(async (req, res, next) => {
+    const user = req.user || {}
     const { taskId } = req.params;
+    const task = await taskService.getTaskById(taskId);
+    if (isEmpty(task)) {
+        return res.status(httpStatus.NOT_FOUND).json({
+            message: "Task not found.",
+            data: null
+        })
+    }
+    /**
+     * NOTE: As of now, only two person can delete the task
+     * 1. User who created the task
+     * 2. Admin
+     * If we don't want this type of functionality, we can comment below condition
+     */
+    if (task?.createdBy?.toString() != user?.id && user.role != roles.ADMIN) {
+        return res.status(httpStatus.BAD_REQUEST).json({
+            message: "Updation not allowed.",
+            data: null
+        })
+    }
     await taskService.deleteTask(taskId);
     return res.status(httpStatus.OK).json({
         message: "success",
@@ -46,7 +67,18 @@ const deleteTask = catchAsync(async (req, res, next) => {
 
 const getAllTasks = catchAsync(async (req, res) => {
     const options = pick(req.query, ['limit', 'page']);
-    const filter = pick(req.query, ['search']);
+    const filter = pick(req.query, ['search', "assignedTo", "status", "priority", "createdBy"]);
+    if (filter.assignedTo) {
+        filter.assignedTo = new mongoose.Types.ObjectId(filter.assignedTo)
+    }
+    if (filter.createdBy) {
+        filter.createdBy = new mongoose.Types.ObjectId(filter.createdBy)
+    }
+
+    if (filter.search) {
+        filter.title = { $regex: filter.search, $options: 'i' };
+        delete filter.search
+    }
     const tasks = await taskService.getAllTasks(filter, options);
     return res.status(200).json({
         message: "success",
@@ -65,6 +97,12 @@ const updateTask = catchAsync(async (req, res) => {
             data: null
         })
     }
+    /**
+     * NOTE: As of now, only two person can edit the task
+     * 1. User who created the task
+     * 2. Admin
+     * If we don't want this type of functionality, we can comment below condition
+     */
     if (task?.createdBy?.toString() != user?.id && user.role != roles.ADMIN) {
         return res.status(httpStatus.BAD_REQUEST).json({
             message: "Updation not allowed.",
